Add outputs field and refactor utils
